Remove some assumptions that 'time_t' is 'long' or 'unsigned long'.
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@27446 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_trace.c
@@ -397,7 +397,7 @@ report_this_status(struct Client *source_p, struct Client *target_p)
 		sendto_one_numeric(source_p, RPL_TRACEUNKNOWN,
 				   form_str(RPL_TRACEUNKNOWN),
 				   class_name, name, ip,
-				   rb_time() - target_p->localClient->firsttime);
+				   (unsigned long)(rb_time() - target_p->localClient->firsttime));
 		cnt++;
 		break;
